view: make setWallpaperPosition() oneway
It is possible that live wallpaper can sleep or too busy
when setWallpaperPosition is called from the launcher.
This can block the launcher process and make launcher to
operate slowly as well. One way to fix this problem is
making setWallpaperPostion() oneway so the launcher
process does not wait for live-wallpaper process

Change-Id: Ifd98ec89945f554d5c53590aa264fb452cc77f23