Fix for "volume buttons change two volume steps per click" bug
This bug happens on all devices since CM9 or CM10 times.

To reproduce:
1) disable Settings/Buttons/Playback control
2) start music playback
3) turn off display
4) click on the volume buttons

When "Playback control" is enabled, the volume changes on the button up
event of short clicks only. When it is disabled, it changes on both
button down and up events. This fix cleans up the code and makes the
volume change only on button down when "Playback control" is disabled.

JIRA: CYAN-3202
Change-Id: If97218e35a8204ef602155f22b7eaaa5e13fb93c

Conflicts:
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java